+++ doannotation_test.go
@@ -64,23 +64,24 @@ func TestDoAnnotation_GetNamed(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments map[string]interface{}
+		arguments map[string]any
 	}
 	tests := []struct {
 		name string
 		args args
 		want string
 	}{
-		{"single", args{name: "k1", arguments: map[string]interface{}{"value": "v1"}}, "v1"},
-		{"multiple", args{name: "k2", arguments: map[string]interface{}{"value": "v2", "num": 2}}, "v2,2"},
-		{"types", args{name: "k_types", arguments: map[string]interface{}{"string": "x", "integer": 1, "float": 2.2, "boolean": true}}, "x,1,2.2,true"},
-		{"person", args{name: "person", arguments: map[string]interface{}{"name": "Saleh", "age": 38, "language": "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
+		{"single", args{name: "k1", arguments: map[string]any{"value": "v1"}}, "v1"},
+		{"multiple", args{name: "k2", arguments: map[string]any{"value": "v2", "num": 2}}, "v2,2"},
+		{"types", args{name: "k_types", arguments: map[string]any{"string": "x", "integer": 1, "float": 2.2, "boolean": true}}, "x,1,2.2,true"},
+		{"person", args{name: "person", arguments: map[string]any{"name": "Saleh", "age": 38, "language": "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
 		{"nil annotation arguments", args{name: "k1", arguments: nil}, "{{value}}"},
-		{"empty annotation argument", args{name: "k1", arguments: map[string]interface{}{}}, "{{value}}"},
-		{"empty annotation", args{name: "k3", arguments: map[string]interface{}{}}, "v3"},
-		{"empty value", args{name: "k4", arguments: map[string]interface{}{"value": "v1"}}, internal.Empty},
-		{"absent annotation", args{name: "k1", arguments: map[string]interface{}{key_NOTFOUND: "v1"}}, "{{value}}"},
-		{"notfound", args{name: key_NOTFOUND, arguments: map[string]interface{}{"value": "v1"}}, internal.Empty},
+		{"empty annotation argument", args{name: "k1", arguments: map[string]any{}}, "{{value}}"},
+		{"empty annotation", args{name: "k3", arguments: map[string]any{}}, "v3"},
+		{"empty value", args{name: "k4", arguments: map[string]any{"value": "v1"}}, internal.Empty},
+		{"absent annotation", args{name: "k1", arguments: map[string]any{key_NOTFOUND: "v1"}}, "{{value}}"},
+		{"few annotations", args{name: "k2", arguments: map[string]any{"value": "v1"}}, "v1,{{num}}"},
+		{"notfound", args{name: key_NOTFOUND, arguments: map[string]any{"value": "v1"}}, internal.Empty},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -106,7 +107,7 @@ func TestDoAnnotation_FindNamed(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments map[string]interface{}
+		arguments map[string]any
 	}
 	tests := []struct {
 		name  string
@@ -114,15 +115,16 @@ func TestDoAnnotation_FindNamed(t *testing.T) {
 		want  string
 		found bool
 	}{
-		{"single", args{name: "k1", arguments: map[string]interface{}{"value": "v1"}}, "v1", true},
-		{"multiple", args{name: "k2", arguments: map[string]interface{}{"value": "v2", "num": 2}}, "v2,2", true},
-		{"types", args{name: "k_types", arguments: map[string]interface{}{"string": "x", "integer": 1, "float": 2.2, "boolean": true}}, "x,1,2.2,true", true},
+		{"single", args{name: "k1", arguments: map[string]any{"value": "v1"}}, "v1", true},
+		{"multiple", args{name: "k2", arguments: map[string]any{"value": "v2", "num": 2}}, "v2,2", true},
+		{"types", args{name: "k_types", arguments: map[string]any{"string": "x", "integer": 1, "float": 2.2, "boolean": true}}, "x,1,2.2,true", true},
 		{"nil annotation arguments", args{name: "k1", arguments: nil}, "{{value}}", true},
-		{"empty annotation argument", args{name: "k1", arguments: map[string]interface{}{}}, "{{value}}", true},
-		{"empty annotation", args{name: "k3", arguments: map[string]interface{}{}}, "v3", true},
-		{"empty value", args{name: "k4", arguments: map[string]interface{}{"value": "v1"}}, internal.Empty, true},
-		{"absent annotation", args{name: "k1", arguments: map[string]interface{}{key_NOTFOUND: "v1"}}, "{{value}}", true},
-		{"notfound", args{name: key_NOTFOUND, arguments: map[string]interface{}{"value": "v1"}}, internal.Empty, false},
+		{"empty annotation argument", args{name: "k1", arguments: map[string]any{}}, "{{value}}", true},
+		{"empty annotation", args{name: "k3", arguments: map[string]any{}}, "v3", true},
+		{"empty value", args{name: "k4", arguments: map[string]any{"value": "v1"}}, internal.Empty, true},
+		{"absent annotation", args{name: "k1", arguments: map[string]any{key_NOTFOUND: "v1"}}, "{{value}}", true},
+		{"few annotations", args{name: "k2", arguments: map[string]any{"value": "v1"}}, "v1,{{num}}", true},
+		{"notfound", args{name: key_NOTFOUND, arguments: map[string]any{"value": "v1"}}, internal.Empty, false},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -152,19 +154,21 @@ func TestDoAnnotation_GetIndexed(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments []interface{}
+		arguments []any
 	}
 	tests := []struct {
 		name string
 		args args
 		want string
 	}{
-		{"single", args{name: "kindexed1", arguments: []interface{}{"v1"}}, "v1"},
-		{"multiple", args{name: "kindexed2", arguments: []interface{}{"1", 2}}, "1,2"},
-		{"types", args{name: "kindexed_types", arguments: []interface{}{"x", 1, 2.2, true}}, "x,1,2.2,true"},
-		{"disordered", args{name: "kindexed2_disordered", arguments: []interface{}{"1", 2, "3", 4}}, "2,1,4,3"},
-		{"over index", args{name: "kindexed1", arguments: []interface{}{"1", "2", "3"}}, "1"},
-		{"person index", args{name: "personindexed", arguments: []interface{}{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
+		{"single", args{name: "kindexed1", arguments: []any{"v1"}}, "v1"},
+		{"multiple", args{name: "kindexed2", arguments: []any{"1", 2}}, "1,2"},
+		{"types", args{name: "kindexed_types", arguments: []any{"x", 1, 2.2, true}}, "x,1,2.2,true"},
+		{"disordered", args{name: "kindexed2_disordered", arguments: []any{"1", 2, "3", 4}}, "2,1,4,3"},
+		{"over index", args{name: "kindexed1", arguments: []any{"1", "2", "3"}}, "1"},
+		{"few index", args{name: "kindexed2", arguments: []any{"1"}}, "1,{{2}}"},
+		{"person index", args{name: "personindexed", arguments: []any{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
+		{"nonindex", args{name: "kindexed_nonindex", arguments: []any{"v1"}}, "{{value}}"},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -190,7 +194,7 @@ func TestDoAnnotation_FindIndexed(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments []interface{}
+		arguments []any
 	}
 	tests := []struct {
 		name  string
@@ -198,12 +202,14 @@ func TestDoAnnotation_FindIndexed(t *testing.T) {
 		want  string
 		found bool
 	}{
-		{"single", args{name: "kindexed1", arguments: []interface{}{"v1"}}, "v1", true},
-		{"multiple", args{name: "kindexed2", arguments: []interface{}{"1", 2}}, "1,2", true},
-		{"types", args{name: "kindexed_types", arguments: []interface{}{"x", 1, 2.2, true}}, "x,1,2.2,true", true},
-		{"disordered", args{name: "kindexed2_disordered", arguments: []interface{}{"1", 2, "3", 4}}, "2,1,4,3", true},
-		{"over index", args{name: "kindexed1", arguments: []interface{}{"1", "2", "3"}}, "1", true},
-		{"person index", args{name: "personindexed", arguments: []interface{}{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer.", true},
+		{"single", args{name: "kindexed1", arguments: []any{"v1"}}, "v1", true},
+		{"multiple", args{name: "kindexed2", arguments: []any{"1", 2}}, "1,2", true},
+		{"types", args{name: "kindexed_types", arguments: []any{"x", 1, 2.2, true}}, "x,1,2.2,true", true},
+		{"disordered", args{name: "kindexed2_disordered", arguments: []any{"1", 2, "3", 4}}, "2,1,4,3", true},
+		{"over index", args{name: "kindexed1", arguments: []any{"1", "2", "3"}}, "1", true},
+		{"few index", args{name: "kindexed2", arguments: []any{"1"}}, "1,{{2}}", true},
+		{"person index", args{name: "personindexed", arguments: []any{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer.", true},
+		{"nonindex", args{name: "kindexed_nonindex", arguments: []any{"v1"}}, "{{value}}", true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -233,17 +239,17 @@ func TestDoAnnotation_GetFormatted(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments []interface{}
+		arguments []any
 	}
 	tests := []struct {
 		name string
 		args args
 		want string
 	}{
-		{"single", args{name: "kformatted1", arguments: []interface{}{"v1"}}, "v1"},
-		{"multiple", args{name: "kformatted2", arguments: []interface{}{"1", 2}}, "1,2"},
-		{"types", args{name: "kformatted_types", arguments: []interface{}{"x", 1, 2.2, true}}, "x,1,2.200000,true"},
-		{"person index", args{name: "personformatted", arguments: []interface{}{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
+		{"single", args{name: "kformatted1", arguments: []any{"v1"}}, "v1"},
+		{"multiple", args{name: "kformatted2", arguments: []any{"1", 2}}, "1,2"},
+		{"types", args{name: "kformatted_types", arguments: []any{"x", 1, 2.2, true}}, "x,1,2.200000,true"},
+		{"person", args{name: "personformatted", arguments: []any{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer."},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -269,7 +275,7 @@ func TestDoAnnotation_FindFormatted(t *testing.T) {
 	}
 	type args struct {
 		name      string
-		arguments []interface{}
+		arguments []any
 	}
 	tests := []struct {
 		name  string
@@ -277,12 +283,12 @@ func TestDoAnnotation_FindFormatted(t *testing.T) {
 		want  string
 		found bool
 	}{
-		{"single", args{name: "kformatted1", arguments: []interface{}{"v1"}}, "v1", true},
-		{"multiple", args{name: "kformatted2", arguments: []interface{}{"1", 2}}, "1,2", true},
-		{"types", args{name: "kformatted_types", arguments: []interface{}{"x", 1, 2.2, true}}, "x,1,2.200000,true", true},
-		{"person index", args{name: "personformatted", arguments: []interface{}{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer.", true},
-		{"notfound", args{name: key_NOTFOUND, arguments: []interface{}{}}, internal.Empty, false},
-		{"empty", args{name: "k4", arguments: []interface{}{}}, internal.Empty, true},
+		{"single", args{name: "kformatted1", arguments: []any{"v1"}}, "v1", true},
+		{"multiple", args{name: "kformatted2", arguments: []any{"1", 2}}, "1,2", true},
+		{"types", args{name: "kformatted_types", arguments: []any{"x", 1, 2.2, true}}, "x,1,2.200000,true", true},
+		{"person", args{name: "personformatted", arguments: []any{"Saleh", 38, "Golang"}}, "Hi, my name is Saleh, I'm 38 years old and a Golang developer.", true},
+		{"notfound", args{name: key_NOTFOUND, arguments: []any{}}, internal.Empty, false},
+		{"empty", args{name: "k4", arguments: []any{}}, internal.Empty, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -312,7 +318,7 @@ func BenchmarkDoAnnotationNamed(b *testing.B) {
 	}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		w.FindNamed("k1", map[string]interface{}{
+		w.FindNamed("k1", map[string]any{
 			"first":  1,
 			"second": 2.22,
 			"three":  "three",